ELEC-198: Moved CAN tests into their own file
